Add guard on token delete. Closes #762. (#763)

+++ quipucords/api/user/view.py
@@ -47,7 +47,9 @@ class UserViewSet(viewsets.GenericViewSet):
         instance = request.user
         logout(request)
         token = Token.objects.filter(user=instance).first()
-        token.delete()
-        token = Token.objects.create(user=instance)
+        if token:
+            token.delete()
+        if instance:
+            token = Token.objects.create(user=instance)
 
         return Response()
